Add files via upload
This is the sample showing how to present data.
You have to convert the .pdf file into a .csv. The column A corresponds to the question code in the questionnaire.
For P1, P4, P7, P10 and P13 : 1 stands for a YES, 2 for a NO.
For all others items, only report the indicated numbers (no need to fill with 0).